### 2018-03-25 #206 @rubenarslan
presubmission inquiry: codebook
Presubmission inquiry following [Twitter contact](https://twitter.com/ma_salmon/status/977192303751323649) with @maelle

Generate reproducible codebooks from data and metadata. The package relies on rmarkdown partials to generate HTML, PDF, and Word documents. Codebooks are also available as tables (CSV, Excel, etc.). HTML codebooks include machine-readable JSON-LD metadata by default. Metadata (var/value/missing labels) as imported by haven or rio is used.

```
Package: codebook
Title: Automatic Codebooks from Survey Metadata Encoded in Attributes
Description: Easily automate the following tasks to describe data frames:
  computing reliabilities (internal consistencies, retest, multilevel) for psychological scales,
  summarise the distributions of scales and items graphically and using descriptive statistics,
  combine this information with metadata (such as item labels and labelled values) that is derived from R attributes.
  To do so, the package relies on 'rmarkdown' partials, so you can generate HTML, PDF, and Word documents. Codebooks
  are also available as tables (CSV, Excel, etc.).
Version: 0.5.8.9000
Authors@R: person("Ruben", "Arslan", email = "ruben.arslan@gmail.com", role = c("aut", "cre"))
		graphics,
    psych,
    likert,
    pander,
    skimr,
    DT,
    future,
    haven,
    mice,
    purrr
  lme4,
  roxygen2
```

https://github.com/rubenarslan/codebook

data deposition, because the package helps generate human and machine-readable codebooks for datasets.

     - People who want to share data or share information about data. Probably mainly social scientists who have smaller amounts of data to share and wouldn't be willing to learn about other, more complex approaches.
     - Users who have received data and want to learn about the dataset, get an overview of included variables, together with metadata.
     - The scientific use here would be to get more metadata out in the open (both in the sense of putting in on the web and using open, standard formats for metadata like JSON-LD). Scientific applications would then be better discovery and interoperability of small datasets (for which the codebook pkg helps only with a small, but necessary preliminary step).

     - Yes
          - [dataMaid](https://github.com/ekstroem/dataMaid) has a codebook function which also uses rmarkdown. I'm in contact with them, and they plan to continue to focus more on error detection. We might collaborate on `codebook`. Features that `codebook` has that `dataMaid` doesn't have: reliability computation, tabular metadata output with `skimr` summaries, machine-readable JSON-LD metadata, embeddable rmarkdown partials.
          - [codebookr](https://github.com/petebaker/codebookr) appears no longer to be under active development and focuses more on extracting structured metadata.
          - [memisc](http://github.com/melff/memisc/) has a codebook function. It focuses on printing codebooks to the console, but also generates HTML. It doesn't generate graphs, or machine-readable metadata and doesn't use the rmarkdown partial approach, which I consider more powerful. Also, memisc appears to be unable to deal with datetimes and labelled vectors.
          - [DDIwR](https://cran.r-project.org/web/packages/DDIwR/index.html) focuses on importing and exporting DDI. I am in contact with author, we'll probably make it possible to generate & embed DDI when generating codebooks.
          - [r2ddi](https://github.com/ddionrails/r2ddi/) is inactive (personal communication: dev. switched to python)
