### 2018-04-04 613d5b7
Variables with only missings don't cause errors.
