+++ man/codebook.Rd
@@ -5,8 +5,9 @@
 \title{Generate rmarkdown codebook}
 \usage{
 codebook(results, reliabilities = NULL, survey_repetition = c("auto",
-  "single", "repeated_once", "repeated_many"), missingness_report = TRUE,
-  metadata_table = TRUE, metadata_json = TRUE, indent = "#")
+  "single", "repeated_once", "repeated_many"), survey_overview = TRUE,
+  missingness_report = TRUE, metadata_table = TRUE, metadata_json = TRUE,
+  indent = "#")
 }
 \arguments{
 \item{results}{a data frame, ideally with attributes set on variables}
@@ -14,6 +15,8 @@ codebook(results, reliabilities = NULL, survey_repetition = c("auto",
 \item{reliabilities}{a named list with one entry per scale and one or several printable reliability computations for this scale. if NULL, computed on-the-fly using compute_reliabilities}
 
 \item{survey_repetition}{defaults to "auto" which is to try to determine the level of repetition from the "session" and "created" variables. Other values are: single, repeated_once, repeated_many}
+
+\item{survey_overview}{whether to print an overview of survey entries, durations (depends on presence of columns session, created, modified, ended, expired)}
 
 \item{missingness_report}{whether to print a missingness report. Turn off if this gets too complicated and you need a custom solution (e.g. in case of random missings).}
 
