### 2018-04-05 8cc111d
escape html chars in various places
