+++ tests/codebook_gss.Rmd
@@ -1,0 +1,44 @@
+---
+title: "Codebook General Social Survey (GSS)"
+output:
+  html_document:
+    toc: true
+    toc_depth: 4
+    toc_float: true
+    code_folding: 'hide'
+    self_contained: false
+---
+
+```{r}
+knitr::opts_chunk$set(warning = FALSE, message = FALSE, error = TRUE)
+pander::panderOptions("table.split.table", Inf)
+```
+
+
+## Load
+```{r echo=FALSE}
+```
+```{r}
+library(codebook)
+results = haven::read_dta("~/research/gss/GSS_stata/GSS7216_R3.DTA")
+# results = results[, c(1:10, which(names(results)=="bthgrp3f"))]
+results <- results
+
+```
+
+```{r}
+# dont cache this chunk
+library(ggplot2)
+theme_set(theme_bw())
+pander::panderOptions("table.split.table", Inf)
+options(device = function(file, width = 7, height = 7, ...) {
+  cairo_pdf(tempfile(), width = width, height = height, ...)
+})
+knitr::opts_chunk$set(echo = FALSE, dev = "CairoPNG")
+library(future)
+plan(multicore(workers = 3))
+```
+```{r}
+codebook(results)
+```
+
