### 2018-04-05 87717dc
- fix dependencies - update pkgdown
