+++ tests/testthat/test_codebook.R
@@ -11,9 +11,9 @@ test_that("codebook generation", {
   dir <- tempdir()
   setwd(dir)
   on.exit(setwd(wd))
-  expect_silent(md <- codebook(bfi))
+  expect_silent(md <- codebook(bfi, metadata_table = FALSE))
   figs <- list.files(paste0(dir, "/figure"))
-  expect_equal(length(figs), 8)
+  expect_equal(length(figs), 7)
   expect_match(md, "Scale: BFIK_neuro")
   expect_match(md, "Scale: BFIK_consc")
   expect_match(md, "Missings per variable")
@@ -40,9 +40,10 @@ test_that("codebook generation without formr", {
   setwd(dir)
   on.exit(setwd(wd))
   expect_silent(md <- codebook(bfi, survey_repetition = "single",
-                               missingness_report = FALSE))
+                               missingness_report = FALSE,
+                               metadata_table = FALSE))
   figs <- list.files(paste0(dir, "/figure"))
-  expect_equal(length(figs), 6)
+  expect_equal(length(figs), 5)
   expect_match(md, "Scale: BFIK_neuro")
   expect_match(md, "Scale: BFIK_consc")
   expect_match(md, "application/ld\\+json")
