### 2018-04-06 4904ef6
- add aggregate_and_document_scale re #1 - update docs
