+++ NAMESPACE
@@ -5,6 +5,7 @@ S3method(summary,labelled)
 S3method(summary,labelled_spss)
 S3method(zap_attributes,data.frame)
 S3method(zap_attributes,default)
+export(aggregate_and_document_scale)
 export(asis_knit_child)
 export(codebook)
 export(codebook_component_scale)
