### 2018-04-06 de24596
- closes #14 by using rio - include rmd example in readme
