### 2018-04-08 b40c1dc
better factor plots
