+++ tests/codebook_example_items_only.Rmd
@@ -1,0 +1,37 @@
+---
+title: "Codebook example formr"
+author: "Ruben Arslan"
+date: "`r Sys.Date()`"
+output:
+  html_document:
+    toc: true
+    toc_depth: 4
+    toc_float: true
+    code_folding: 'hide'
+    self_contained: true
+---
+
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+
+```{r}
+knitr::opts_chunk$set(warning = TRUE, message = TRUE, echo = FALSE, error = FALSE)
+pander::panderOptions("table.split.table", Inf)
+library(codebook)
+data("bfi", package = 'codebook')
+bfi$BFIK_agree <- NULL
+ggplot2::theme_set(ggplot2::theme_bw())
+pander::panderOptions("table.split.table", Inf)
+bfi$age <- rpois(nrow(bfi), 30)
+```
+
+```{r cb}
+codebook(bfi)
+```
+
