+++ tests/codebook_example_factor.Rmd
@@ -1,0 +1,43 @@
+---
+title: "Codebook example formr"
+author: "Ruben Arslan"
+date: "`r Sys.Date()`"
+output:
+  html_document:
+    toc: true
+    toc_depth: 4
+    toc_float: true
+    code_folding: 'hide'
+    self_contained: true
+vignette: >
+  %\VignetteIndexEntry{Codebook example formr}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+
+```{r}
+knitr::opts_chunk$set(warning = TRUE, message = TRUE, echo = FALSE, error = TRUE)
+pander::panderOptions("table.split.table", Inf)
+library(codebook)
+library(dplyr)
+data("bfi", package = 'codebook')
+bfi = bfi %>% haven::zap_missing() %>% haven::as_factor() %>% 
+  select(-BFIK_extra, -BFIK_neuro, -BFIK_agree, -BFIK_open, -BFIK_consc)
+ggplot2::theme_set(ggplot2::theme_bw())
+pander::panderOptions("table.split.table", Inf)
+bfi$age <- rpois(nrow(bfi), 30)
+```
+
+```{r cb}
+codebook(bfi)
+```
+
