### 2018-04-19 #206 @noamross
Sorry for the delay, @rubenarslan. This got lost in the shuffle as we rotated editors.  Ultimately, we decided that **codebook** doesn't fall into our scope.  The data deposition category is primarily for uploading data to scientific repositories and generating appropriate metadata for that purpose.  **codebook** falls more into the category of literate programming or data exploration tools to produce human-readable outputs from data, which we don't support anymore because we've decided that that we can't really judge a 'best' way to look at outputs.

Of course, being out-of-scope for us doesn't mean the package isn't useful!  I think this kind of thing is very handy.  We encourage you to submit to other options like JOSS, where I believe this would be in-scope.
