### 2018-04-19 #206 @rubenarslan
@noamross To me the human-readable part is more a means to the end of getting machine-readable and Google-able metadata out there and into repositories. But it's definitely a more minor part of what the package actually does. I was hoping to benefit from the metadata expertise of some of the reviewers here, but I guess/hope some of the same people review for JOSS.
