### 2018-04-19 #206 @noamross
Yes, there is overlap in both reviewers and editors between us and JOSS. We'll tell the JOSS editors to look out for the package to make sure the right folks handle it.
